+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Transfer.java
@@ -3,11 +3,11 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class Transfer {
-    private long transfer_id;
+    private int transfer_id;
     private int transfer_type_id;
     private int transfer_status_id;
-    private long account_to;
-    private long account_from;
+    private int account_to;
+    private int account_from;
     private BigDecimal amount;
     private String userFrom;
     private String  userTo;
@@ -33,7 +33,7 @@ public class Transfer {
         this.transfer_status_desc = transfer_status_desc;
     }
 
-    public Transfer(long transfer_id, int transfer_type_id, int transfer_status_id, long account_to, long account_from, BigDecimal amount) {
+    public Transfer(int transfer_id, int transfer_type_id, int transfer_status_id, int account_to, int account_from, BigDecimal amount) {
         this.transfer_id = transfer_id;
         this.transfer_type_id = transfer_type_id;
         this.transfer_status_id = transfer_status_id;
@@ -42,11 +42,11 @@ public class Transfer {
         this.amount = amount;
     }
 
-    public long getTransfer_id() {
+    public int getTransfer_id() {
         return transfer_id;
     }
 
-    public void setTransfer_id(long transfer_id) {
+    public void setTransfer_id(int transfer_id) {
         this.transfer_id = transfer_id;
     }
 
@@ -66,19 +66,19 @@ public class Transfer {
         this.transfer_status_id = transfer_status_id;
     }
 
-    public long getAccount_to() {
+    public int getAccount_to() {
         return account_to;
     }
 
-    public void setAccount_to(long account_to) {
+    public void setAccount_to(int account_to) {
         this.account_to = account_to;
     }
 
-    public long getAccount_from() {
+    public int getAccount_from() {
         return account_from;
     }
 
-    public void setAccount_from(long account_from) {
+    public void setAccount_from(int account_from) {
         this.account_from = account_from;
     }
